backend/requirements/base.txt

# Bleeding edge Django
django==1.10.2
django-environ==0.4.0
django-allauth==0.25.2